refactor(tb6612fng): Rework tb6612fng_set_duty_cycle() to use timer_pwm_set_compare_value() instead of timer_pwm_set_duty_cycle() and update changes everywhere

+++ Devices/dc_motor/dc_motor.c
@@ -10,7 +10,7 @@ DC_MOTOR_Status_t dc_motor_init(DC_MOTOR_t *motor)
 	return DC_MOTOR_OK;
 }
 
-DC_MOTOR_Status_t dc_motor_set_speed(DC_MOTOR_t *motor, uint8_t speed)
+DC_MOTOR_Status_t dc_motor_set_speed(DC_MOTOR_t *motor, uint16_t speed)
 {
 	if (tb6612fng_set_duty_cycle(motor->motor_driver, motor->driver_channel, speed) != TB6612FNG_OK) return DC_MOTOR_ERR;
 
@@ -24,7 +24,7 @@ DC_MOTOR_Status_t dc_motor_set_direction(DC_MOTOR_t *motor, TB6612FNG_Direction_
 	return DC_MOTOR_OK;
 }
 
-DC_MOTOR_Status_t dc_motor_set_speed_and_direction(DC_MOTOR_t *motor, TB6612FNG_Direction_t direction, uint8_t speed)
+DC_MOTOR_Status_t dc_motor_set_speed_and_direction(DC_MOTOR_t *motor, TB6612FNG_Direction_t direction, uint16_t speed)
 {
 	if (dc_motor_set_direction(motor, direction) != DC_MOTOR_OK) return DC_MOTOR_ERR;
 	if (dc_motor_set_speed(motor, speed) != DC_MOTOR_OK) return DC_MOTOR_ERR;
